Claim people action!
So, now users can go to an event page, search their names and take it,
so they can later find out who they'll be gifting to.

Just had some issues with streaming forms with hotwire, so, easy
solution was to use stimullus to make async requests and left rails to
do the magic by broadcasting an event on updates-commit, so the
claimed-people-conter magically keeps updated!