Reveal your matches :fire:
Now, when a user pick their name from the list, a link appears to
reveal who they'll be gifting to.

Since a user can take more than one card, for cases like.. a family or
a couple that is participating together in the event, instead of having
to take one by one in different sessions, they can do it in one shot,
so this page supports displaying a list of matches as well.